x86_64: Rename from x86 for consistency
This will break the API but makes it consistent with different
architectures (so it matches the kernel arch name that it was generated
from).
